% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_func.R
\name{mround}
\alias{mround}
\title{query of AWS owenslake database}
\usage{
mround(x, base)
}
\arguments{
\item{x.}{Number to round.}

\item{base.}{Base to round to.}
}
\description{
query of AWS owenslake database
}

