% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_func.R
\name{query_owens}
\alias{query_owens}
\title{query of AWS owenslake database}
\usage{
query_owens(query)
}
\arguments{
\item{query}{String. Postgres query for database.}
}
\description{
query of AWS owenslake database
}

