% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_func.R
\name{shoreline_roses}
\alias{shoreline_roses}
\title{Build map image of dust roses for shoreline TEOM stations}
\usage{
shoreline_roses(day = Sys.Date() - 1, path = tempdir())
}
\arguments{
\item{day}{Date. Date for which to build dust roses.}

\item{path}{String. Path to image file.}
}
\value{
Text string denoting location of plot image file.
}
\description{
Build map image of dust roses for shoreline TEOM stations
}

